requirements.txt
wheel
pyserial==3.5
PyYAML==6.0.1
python-socketio
eventlet
websocket-client
requests
